Move import pyfits stuff to galsim/_pyfits.py, so it can be delayed until actually needed.
