+++ galsim/fits.py
@@ -25,7 +25,6 @@ routines for handling multiple Images.
 
 import os
 import galsim
-from galsim import pyfits, pyfits_version, pyfits_str
 
 
 ##############################################################################################
@@ -36,6 +35,7 @@ from galsim import pyfits, pyfits_version, pyfits_str
 ##############################################################################################
  
 def _parse_compression(compression, file_name):
+    from galsim._pyfits import pyfits, pyfits_version
     file_compress = None
     pyfits_compress = None
     if compression == 'rice' or compression == 'RICE_1': pyfits_compress = 'RICE_1'
@@ -71,6 +71,7 @@ class _ReadFile:
         # cf. http://bugs.python.org/issue7471
         import subprocess
         from cStringIO import StringIO
+        from galsim._pyfits import pyfits
         # We use gunzip -c rather than zcat, since the latter is sometimes called gzcat
         # (with zcat being a symlink to uncompress instead).
         p = subprocess.Popen(["gunzip", "-c", file], stdout=subprocess.PIPE, close_fds=True)
@@ -81,6 +82,7 @@ class _ReadFile:
 
     def gzip_in_mem(self, file):
         import gzip
+        from galsim._pyfits import pyfits
         fin = gzip.open(file, 'rb')
         hdu_list = pyfits.open(fin, 'readonly')
         # Sometimes this doesn't work.  The symptoms may be that this raises an
@@ -93,6 +95,7 @@ class _ReadFile:
         return hdu_list, fin
 
     def pyfits_open(self, file):
+        from galsim._pyfits import pyfits
         # This usually works, although pyfits internally may (depending on the version)
         # use a temporary file, which is why we prefer the above in-memory code if it works.
         # For some versions of pyfits, this is actually the same as the in_mem version.
@@ -101,6 +104,7 @@ class _ReadFile:
 
     def gzip_tmp(self, file):
         import gzip
+        from galsim._pyfits import pyfits
         # Finally, just in case, if everything else failed, here is an implementation that 
         # should always work.
         fin = gzip.open(file, 'rb')
@@ -117,6 +121,7 @@ class _ReadFile:
     def bunzip2_call(self, file):
         import subprocess
         from cStringIO import StringIO
+        from galsim._pyfits import pyfits
         p = subprocess.Popen(["bunzip2", "-c", file], stdout=subprocess.PIPE, close_fds=True)
         fin = StringIO(p.communicate()[0])
         assert p.returncode == 0 
@@ -125,6 +130,7 @@ class _ReadFile:
 
     def bz2_in_mem(self, file):
         import bz2
+        from galsim._pyfits import pyfits
         # This normally works.  But it might not on old versions of pyfits.
         fin = bz2.BZ2File(file, 'rb')
         hdu_list = pyfits.open(fin, 'readonly')
@@ -136,6 +142,7 @@ class _ReadFile:
 
     def bz2_tmp(self, file):
         import bz2
+        from galsim._pyfits import pyfits
         fin = bz2.BZ2File(file, 'rb')
         data = fin.read()
         tmp = file + '.tmp'
@@ -165,6 +172,7 @@ class _ReadFile:
         self.bz2 = self.bz2_methods[0]
 
     def __call__(self, file, dir, file_compress):
+        from galsim._pyfits import pyfits, pyfits_version
         if dir:
             import os
             file = os.path.join(dir,file)
@@ -337,6 +345,7 @@ class _WriteFile:
 
     def __call__(self, file, dir, hdu_list, clobber, file_compress, pyfits_compress):
         import os
+        from galsim._pyfits import pyfits, pyfits_version
         if dir:
             file = os.path.join(dir,file)
 
@@ -395,6 +404,7 @@ class _WriteFile:
 _write_file = _WriteFile()
 
 def _add_hdu(hdu_list, data, pyfits_compress):
+    from galsim._pyfits import pyfits, pyfits_version
     if pyfits_compress:
         if len(hdu_list) == 0:
             hdu_list.append(pyfits.PrimaryHDU())  # Need a blank PrimaryHDU
@@ -414,6 +424,7 @@ def _add_hdu(hdu_list, data, pyfits_compress):
 def _check_hdu(hdu, pyfits_compress):
     """Check that an input `hdu` is valid
     """
+    from galsim._pyfits import pyfits
     if pyfits_compress:
         if not isinstance(hdu, pyfits.CompImageHDU):
             if isinstance(hdu, pyfits.BinTableHDU):
@@ -431,6 +442,7 @@ def _check_hdu(hdu, pyfits_compress):
 
 
 def _get_hdu(hdu_list, hdu, pyfits_compress):
+    from galsim._pyfits import pyfits
     if isinstance(hdu_list, pyfits.HDUList):
         # Note: Nothing special needs to be done when reading a compressed hdu.
         # However, such compressed hdu's may not be the PrimaryHDU, so if we think we are
@@ -514,7 +526,8 @@ def write(image, file_name=None, dir=None, hdu_list=None, clobber=True, compress
                                    otherwise None
                         [default: 'auto']
     """
-  
+    from galsim._pyfits import pyfits
+
     file_compress, pyfits_compress = _parse_compression(compression,file_name)
 
     if file_name and hdu_list is not None:
@@ -555,6 +568,7 @@ def writeMulti(image_list, file_name=None, dir=None, hdu_list=None, clobber=True
     @param clobber      See documentation for this parameter on the galsim.fits.write() method.
     @param compression  See documentation for this parameter on the galsim.fits.write() method.
     """
+    from galsim._pyfits import pyfits
 
     file_compress, pyfits_compress = _parse_compression(compression,file_name)
 
@@ -609,6 +623,7 @@ def writeCube(image_list, file_name=None, dir=None, hdu_list=None, clobber=True,
     @param compression  See documentation for this parameter on the galsim.fits.write() method.
     """
     import numpy
+    from galsim._pyfits import pyfits
 
     file_compress, pyfits_compress = _parse_compression(compression,file_name)
 
@@ -827,6 +842,7 @@ def readMulti(file_name=None, dir=None, hdu_list=None, compression='auto'):
 
     @returns a Python list of Images
     """
+    from galsim._pyfits import pyfits
 
     file_compress, pyfits_compress = _parse_compression(compression,file_name)
 
@@ -1081,6 +1097,7 @@ class FitsHeader(object):
 
     def __init__(self, header=None, file_name=None, dir=None, hdu_list=None, hdu=None,
                  compression='auto', text_file=False):
+        from galsim._pyfits import pyfits
 
         if header and file_name:
             raise TypeError("Cannot provide both file_name and header to FitsHeader")
@@ -1171,12 +1188,14 @@ class FitsHeader(object):
         return self.header.__iter__
 
     def __len__(self):
+        from galsim._pyfits import pyfits_version
         if pyfits_version < '3.1':
             return len(self.header.ascard)
         else:
             return len(self.header)
 
     def __setitem__(self, key, value):
+        from galsim._pyfits import pyfits_version
         self._tag = None
         if pyfits_version < '3.1':
             if isinstance(value, tuple):
@@ -1220,6 +1239,7 @@ class FitsHeader(object):
         return self.header.keys()
 
     def update(self, dict2):
+        from galsim._pyfits import pyfits_version
         self._tag = None
         # dict2 may be a dict or another FitsHeader (or anything that acts like a dict).
         if pyfits_version < '3.1':
@@ -1242,6 +1262,7 @@ class FitsHeader(object):
         @param useblanks    If there are blank entries currently at the end, should they be
                             overwritten with the new entry? [default: True]
         """
+        from galsim._pyfits import pyfits, pyfits_version
         self._tag = None
         if pyfits_version < '3.1':
             self.header.ascardlist().append(pyfits.Card(key, value),
@@ -1250,6 +1271,7 @@ class FitsHeader(object):
             self.header.append((key, value), useblanks=useblanks, bottom=True)
 
     def __repr__(self):
+        from galsim._pyfits import pyfits_str
         if self._tag is None:
             return "galsim.FitsHeader(header=%s.Header().fromstring(%r))"%(
                     pyfits_str,str(self.header))
